Save _v_lang for each field in request instance instead of field instance, if possible.
There was a discrepancy with the current implementation, when dealing with an instance having multiple threads. Since each field instance is shared between all threads, there could be situations, when the call of setLanguage in one thread could influence the outcome of another thread (e.g. like setting a field to a non-primary language, which could yield a faulty value possibly leading to errors). We fix this behaviour by using each thread's request object as data store whenever possible. If no request is present, the previous implementation will be used.